+++ src/DatabaseMigrationTools/DatabaseMigrationTools.csproj
@@ -75,10 +75,10 @@
       <Version>4.1.0-master-2758883</Version>
     </PackageReference>
     <PackageReference Include="NuGet.Services.Configuration">
-      <Version>2.53.0</Version>
+      <Version>2.54.0</Version>
     </PackageReference>
     <PackageReference Include="NuGet.Services.Validation">
-      <Version>2.53.0</Version>
+      <Version>2.54.0</Version>
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
