+++ src/NuGet.Services.DatabaseMigration/NuGet.Services.DatabaseMigration.csproj
@@ -91,7 +91,7 @@
       <Version>4.1.0-master-2602271</Version>
     </PackageReference>
     <PackageReference Include="NuGet.Services.Configuration">
-      <Version>2.53.0</Version>
+      <Version>2.54.0</Version>
     </PackageReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
