+++ tests/NuGetGallery.Core.Facts/NuGetGallery.Core.Facts.csproj
@@ -80,7 +80,7 @@
     <Compile Include="Certificates\CertificateFileFacts.cs" />
     <Compile Include="Cookies\CookieComplianceServiceBaseFacts.cs" />
     <Compile Include="Features\FeatureFlagClientExtensionsFacts.cs" />
-    <Compile Include="Features\FeatureFlagFileStorageServiceFacts.cs" />
+    <Compile Include="Features\EditableFeatureFlagFileStorageServiceFacts.cs" />
     <Compile Include="GitHub\GitHubUsageConfigurationFacts.cs" />
     <Compile Include="Infrastructure\ElmahExceptionFacts.cs" />
     <Compile Include="Extensions\UserExtensionsCoreFacts.cs" />
